Prepare 1.1 Changes (#7)
Convert code into Classes
Remove keptn functions and unused TXT files
Add Readme files
Add Examples
