--- vars/dt_keptnHelpers.groovy
@@ -1,7 +1,0 @@
-def get_timestamp(){
-    DATE_TAG = java.time.LocalDate.now()
-    DATETIME_TAG = java.time.LocalDateTime.now()
-    //echo "${DATETIME_TAG}"
-                
-  return DATETIME_TAG
-}
