--- vars/pushDynatraceConfigurationEvent.txt
@@ -1,7 +1,0 @@
-## pushDynatraceConfigurationEvent
-Used to push a Configuration Changed Event to Dynatrace
-
-### *usage:*
-errorCode=pushDynatraceConfigurationEvent(tagRule)
-if(errorCode > 0)
-  error "Pushing event to Dynatrace failed"
