--- vars/pushDynatraceDeploymentEvent.txt
@@ -1,7 +1,0 @@
-## pushDynatraceDeploymentEvent
-Used to push a Deployment Event to Dynatrace
-
-### *usage:*
-errorCode=pushDynatraceDeploymentEvent(tagRule)
-if(errorCode > 0)
-  error "Pushing event to Dynatrace failed"
